requirements.txt
torch
torchvision
tqdm
numpy
pillow
pycocotools
torch
torchvision
opencv-python